extract model into core module
