+++ core/model.py
@@ -1,0 +1,24 @@
+from keras.models import Sequential
+from keras.layers import Rescaling, Conv2D, MaxPooling2D, Flatten, Dense
+from keras.losses import SparseCategoricalCrossentropy
+
+
+def compile_model(num_classes):
+    model = Sequential([
+        Rescaling(1./255),
+        Conv2D(32, 3, activation='relu'),
+        MaxPooling2D(),
+        Conv2D(32, 3, activation='relu'),
+        MaxPooling2D(),
+        Conv2D(32, 3, activation='relu'),
+        MaxPooling2D(),
+        Flatten(),
+        Dense(128, activation='relu'),
+        Dense(num_classes),
+    ])
+    model.compile(
+        optimizer='adam',
+        loss=SparseCategoricalCrossentropy(from_logits=True),
+        metrics=['accuracy'],
+    )
+    return model
