+++ main.py
@@ -1,32 +1,10 @@
-from keras.models import Sequential
-from keras.layers import Rescaling, Conv2D, MaxPooling2D, Flatten, Dense
-from keras.losses import SparseCategoricalCrossentropy
 import matplotlib.pyplot as plt
 
 from core.loader import load_dataset
+from core.model import compile_model
 from utils.plot import save_plot
 import configs.model as config
 
-
-def compile_model(num_classes):
-    model = Sequential([
-        Rescaling(1./255),
-        Conv2D(32, 3, activation='relu'),
-        MaxPooling2D(),
-        Conv2D(32, 3, activation='relu'),
-        MaxPooling2D(),
-        Conv2D(32, 3, activation='relu'),
-        MaxPooling2D(),
-        Flatten(),
-        Dense(128, activation='relu'),
-        Dense(num_classes),
-    ])
-    model.compile(
-        optimizer='adam',
-        loss=SparseCategoricalCrossentropy(from_logits=True),
-        metrics=['accuracy'],
-    )
-    return model
 
 def plot_history(history):
     fig, ax = plt.subplots()
